% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Data from the Giesikus model}
\format{
A data frame with 1024 rows and 4 columns
\describe{
\item{V1}{Time}
\item{V2}{Strain}
\item{V3}{Rate}
\item{V4}{Stress}
}
}
\usage{
data(mydata)
}
\description{
The data is arranged into four columns: Time (s), Strain (-), Rate (1/s) and Stress (Pa). reflecting the applied strain- control
}
\references{
ppp
}
\keyword{datasets}
