% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_ggplot_v3.r, R/Rpp_plot_v3.r
\name{plotColeCole}
\alias{plotColeCole}
\title{Cole-Cole plot}
\usage{
plotColeCole(Gp_t, Gpp_t, ...)

plotColeCole(Gp_t, Gpp_t, ...)
}
\arguments{
\item{Gp_t}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{Gpp_t}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{...}{parameters of plot()}
}
\value{
{No return value}

{No return value}
}
\description{
create Cole-Cole plot

create Cole-Cole plot
}
\examples{
\donttest{ data(mydata)
df <- rpp_read2(mydata , selected=c(2, 3, 4, 0, 0, 1, 0, 0))
time_wave <- df$raw_time
resp_wave <- data.frame(df$strain,df$strain_rate,df$stress) 
out <- Rpp_num(time_wave, resp_wave , L=1024, k=8, num_mode=1)
Gp_t= out$spp_data_out$Gp_t
Gpp_t= out$spp_data_out$Gpp_t
plotColeCole(Gp_t,Gpp_t)}

}
\references{
Simon A. Rogersa, M. Paul Letting, A sequence of physical processes determined and quantified in large-amplitude oscillatory shear (LAOS): Application to theoretical nonlinear models
Journal of Rheology 56:1, 1-25

Simon A. Rogersa, M. Paul Letting, A sequence of physical processes determined and quantified in large-amplitude oscillatory shear (LAOS): Application to theoretical nonlinear models
Journal of Rheology 56:1, 1-25
}
\author{
Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter

Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter
}
\keyword{Cole-Cole}
\keyword{SPP}
\keyword{plot}
