% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_ggplot_v3.r, R/Rpp_plot_v3.r
\name{plotFft}
\alias{plotFft}
\title{Fourier Harmonic Magnitudes plot}
\usage{
plotFft(ft_amp, fft_resp, spp_params, ...)

plotFft(ft_amp, fft_resp, spp_params, ...)
}
\arguments{
\item{ft_amp}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{fft_resp}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{spp_params}{input parameters used for the fft analysis or numerical differentiation analysis}

\item{...}{parameters of plot()}
}
\value{
{No return value}

{No return value}
}
\description{
create Fourier Harmonic Magnitudes plot

create Fourier Harmonic Magnitudes plot
}
\examples{
\donttest{ data(mydata)
df <- rpp_read2(mydata , selected=c(2, 3, 4, 0, 0, 1, 0, 0))
time_wave <- df$raw_time
resp_wave <- data.frame(df$strain,df$strain_rate,df$stress) 
out <- rpp_fft(time_wave,resp_wave,L=1024,omega=3.16 , M=15,p=1)
ft_amp= out$ft_out$ft_amp
fft_resp= out$ft_out$fft_resp
spp_params= out$spp_params
plotFft(ft_amp,fft_resp,spp_params)}

}
\author{
Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter

Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter
}
