% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_ggplot_v3.r, R/Rpp_plot_v3.r
\name{plotStressTime}
\alias{plotStressTime}
\title{Stress-Time plot}
\usage{
plotStressTime(time_wave_in, stress_in, time_wave, stress)

plotStressTime(time_wave_in, stress_in, time_wave, stress)
}
\arguments{
\item{time_wave_in}{raw time from input data}

\item{stress_in}{stress from input data}

\item{time_wave}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{stress}{from the output matrix from fft analysis or numerical differentiation analysis}
}
\value{
{No return value}

{No return value}
}
\description{
create Stress-Time plot

create Stress-Time plot
}
\examples{
\donttest{ data(mydata)
df <- rpp_read2(mydata , selected=c(2, 3, 4, 0, 0, 1, 0, 0))
time_wave <- df$raw_time
resp_wave <- data.frame(df$strain,df$strain_rate,df$stress) 
out <- Rpp_num(time_wave, resp_wave , L=1024, k=8, num_mode=1)
time_wave_in= out$spp_data_in$time_wave
stress_in= out$spp_data_in$stress
time_wave= out$spp_data_out$time_wave
stress= out$spp_data_out$stress
plotStressTime(time_wave_in,stress_in,time_wave,stress)}

}
\author{
Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter

Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter
}
