% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_ggplot_v3.r, R/Rpp_plot_v3.r
\name{plotTimeStrain}
\alias{plotTimeStrain}
\title{Strain time_wave,strain}
\usage{
plotTimeStrain(time_wave, strain, time_wave_in, strain_in, ...)

plotTimeStrain(time_wave, strain, time_wave_in, strain_in, ...)
}
\arguments{
\item{time_wave}{time from output data}

\item{strain}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{time_wave_in}{time from input data}

\item{strain_in}{from the input matrix from fft analysis or numerical differentiation analysis}

\item{...}{parameters of plot()}
}
\value{
{No return value}

{No return value}
}
\description{
Strain time_wave, strain

Strain time_wave, strain
}
\examples{
\donttest{ data(mydata)
df <- rpp_read2(mydata , selected=c(2, 3, 4, 0, 0, 1, 0, 0))
time_wave <- df$raw_time
resp_wave <- data.frame(df$strain,df$strain_rate,df$stress) 
out <- Rpp_num(time_wave, resp_wave , L=1024, k=8, num_mode=1)
time_wave= out$spp_data_out$time_wave
strain= out$spp_data_out$strain
time_wave_in= out$spp_data_in$time_wave
strain_in= out$spp_data_in$strain
plotTimeStrain(time_wave,strain,time_wave_in,strain_in)}

}
\references{
Simon A. Rogersa, M. Paul Letting, A sequence of physical processes determined and quantified in large-amplitude oscillatory shear (LAOS): Application to theoretical nonlinear models
Journal of Rheology 56:1, 1-25

Simon A. Rogersa, M. Paul Letting, A sequence of physical processes determined and quantified in large-amplitude oscillatory shear (LAOS): Application to theoretical nonlinear models
Journal of Rheology 56:1, 1-25
}
