% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_ggplot_v3.r, R/Rpp_plot_v3.r
\name{plotTimeStress}
\alias{plotTimeStress}
\title{Stress-Time plot}
\usage{
plotTimeStress(time_wave, stress, time_wave_in, strain_rate, ...)

plotTimeStress(time_wave, stress, time_wave_in, strain_rate, ...)
}
\arguments{
\item{time_wave}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{stress}{from the output matrix from fft analysis or numerical differentiation analysis}

\item{time_wave_in}{raw time from input data}

\item{strain_rate}{strain rate from input data}

\item{...}{parameters of plot()}
}
\value{
{No return value}

{No return value}
}
\description{
create Stress-Time plot

create Stress-Time plot
}
\examples{
\donttest{ data(mydata)
df <- rpp_read2(mydata , selected=c(2, 3, 4, 0, 0, 1, 0, 0))
time_wave <- df$raw_time
resp_wave <- data.frame(df$strain,df$strain_rate,df$stress) 
out <- Rpp_num(time_wave, resp_wave , L=1024, k=8, num_mode=1)
time_wave= out$spp_data_out$time_wave
stress= out$spp_data_out$stress
time_wave_in= out$spp_data_in$time_wave
strain_rate= out$spp_data_in$strain_rate
plotTimeStress(time_wave,stress,time_wave_in,strain_rate)}

}
\author{
Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter

Giorgio Luciano and Serena Beretta, based on the Plotting functions created by Simon Rogers Group for Soft Matter
}
