% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_out.r
\name{rpp_out_excel}
\alias{rpp_out_excel}
\title{Export results of the performed SPP analysis in xls format}
\usage{
rpp_out_excel(out, myfilename = "my_models.xlsx")
}
\arguments{
\item{out}{output of the SPP analysis (performed via FFT or Numeric Analysis)}

\item{myfilename}{name of the file where to save results in xls format}
}
\value{
{No return value}
}
\description{
# This function export the output the SPP analysis (performed via FFT or Numeric Analysis) and export it to xls files
}
\author{
Simon Rogers Group for Soft Matter (matlab version), Giorgio Luciano and Serena Berretta (R version)
}
\keyword{export}
\keyword{results}
\keyword{xls}
