% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpp_read.r
\name{rpp_read}
\alias{rpp_read}
\title{Read function}
\usage{
rpp_read(filename, header = TRUE, selected = c(2, 3, 4, 0, 0, 1, 0, 0), ...)
}
\arguments{
\item{filename}{the name of the file to read}

\item{header}{TRUE if colnames are present FALSE if colnames are not present}

\item{selected}{the user should input the number of the columns that represent
strain-smoothed (gamma), strain rate-smoothed (gamma dot), stress smoothed (tau recon), Elast-Stress (FTtau_e), Visco-Stress (FTtau_v), raw time (time), raw stress (tau), raw strain (gamma)
i.e. selected=c(2, 3, 4, 0, 0, 1, 0, 0) means that the second column of your data is the strain rate smoothed, the third column is the stress smoothed, the stress smoothed is the fourth column in the original data,
and finally that we do not have data for the raw stress and raw strain}

\item{...}{parameters of read.csv}
}
\value{
a dataframe with all the columns assigned
}
\description{
This function reads data from the selected file, and assign it to a dataframe
}
\author{
Giorgio Luciano and Serena Berretta, Simon Rogers Group for Soft Matter (matlab version)
}
